refactor: create githubprofile-service
